fix/suppress some warnings (#1109)
* Test then punt on Cython 3.0a8. The vectorcall optimizations weren't noticeable. It can fix the `NPY_NO_DEPRECATED_API` warnings but prints more `tp_print` warnings. (Python 3.8 deprecated the `PyObject.tp_print` field so C++ warns when either initializing _or_ failing to initialize it.)
* Test then punt on NumPy 1.21.0 for now. The SIMD optimizations weren't noticeable. The built-in type stubs didn't help all the PyCharm warnings "Parameter 'initial' unfilled" for ndarray.min() and max().
* Fix 2 mypy errors.
* Suppress `RuntimeWarning: invalid value encountered in subtract` (presumably NaN values) from comparePickles' object_tree diff.
* `np.bool`, `np.int`, and `np.float` are long-deprecated aliases for `bool` etc. Replace them to avoid deprecation warnings in NumPy 1.20+.
* Include Parca analyses in wcm.py help.
* Suppress numpy "divide" and "invalid" warnings in 5 more analyses per nightly anaerobic build.
* Fix/suppress some PyCharm inspection warnings.
* Does anyone use `wholecell/utils/config.py`? It loads a JSON file `default_config.cfg` with the wrong filename extension, causing an inspection warning, and blindly loads it into the global variables dict! It looks safe to delete.
* .gitignore .vscode/ in case anyone wants to try GitHub Copilot ;)

+++ models/ecoli/analysis/multigen/functionalUnits.py
@@ -100,7 +100,7 @@ class Plot(multigenAnalysisPlot.MultigenAnalysisPlot):
 			# Account for bulk molecules
 			bulkMolecules = TableReader(os.path.join(simOutDir, "BulkMolecules"))
 			moleculeIds = bulkMolecules.readAttribute("objectNames")
-			proteinIndexes = np.array([moleculeIds.index(moleculeId) for moleculeId in ids_protein], np.int)
+			proteinIndexes = np.array([moleculeIds.index(moleculeId) for moleculeId in ids_protein], int)
 			proteinCountsBulk = bulkMolecules.readColumn("counts")[:, proteinIndexes]
 			bulkMolecules.close()
